requirements.txt
fastapi[all]
Pillow
python-multipart
torch
transformers
ipykernel